+++ src/main/java/Drink.java
@@ -1,0 +1,6 @@
+public class Drink extends MenuItem{
+    public void printName()
+    {
+        System.out.println(this.name);
+    }
+}
