+++ src/main/java/MenuItem.java
@@ -1,0 +1,5 @@
+public class MenuItem {
+    String name;
+    String type;
+    String[] material;
+}
